refactor: split pii redaction backend modules
