+++ crates/core/src/plugins/pii_redaction/component.rs
@@ -7,34 +7,23 @@ use std::future::Future;
 use std::pin::Pin;
 use std::sync::Arc;
 
-use regex::Regex;
-use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
-use serde_json::{Map, Value as Json};
-use sha2::{Digest, Sha256};
-
-use crate::api::llm::LlmRequest;
-use crate::api::runtime::{LlmSanitizeRequestFn, LlmSanitizeResponseFn, ToolSanitizeFn};
-use crate::codec::anthropic::AnthropicMessagesCodec;
-use crate::codec::openai_chat::OpenAIChatCodec;
-use crate::codec::openai_responses::OpenAIResponsesCodec;
-use crate::codec::request::{ContentPart, MessageContent};
-use crate::codec::response::{AnnotatedLlmResponse, FinishReason, ResponseToolCall};
-use crate::codec::traits::{LlmCodec, LlmResponseCodec};
 use crate::plugin::{
     ConfigDiagnostic, ConfigPolicy, DiagnosticLevel, Plugin, PluginComponentSpec, PluginError,
     PluginRegistrationContext, Result as PluginResult, UnsupportedBehavior, deregister_plugin,
     register_plugin,
 };
+use serde::{Deserialize, Serialize};
+use serde_json::{Map, Value as Json};
 
-#[path = "local.rs"]
-mod local;
-use local::register_local_backend;
-pub use local::{clear_local_backend_provider, register_local_backend_provider};
-
-#[path = "detectors.rs"]
-mod detectors;
-use detectors::{BuiltinDetector, detector_regex_pattern, supported_detector_summary};
+use super::builtin::{
+    CompiledBuiltinBackend, llm_sanitize_request_callback, llm_sanitize_response_callback,
+    tool_sanitize_callback,
+};
+#[cfg(test)]
+pub(crate) use super::builtin::{hex_sha256, mask_text};
+use super::detectors::{detector_regex_pattern, supported_detector_summary};
+use super::local::register_local_backend;
+pub use super::local::{clear_local_backend_provider, register_local_backend_provider};
 
 /// The plugin kind reserved for the built-in privacy component.
 pub const PII_REDACTION_PLUGIN_KIND: &str = "pii_redaction";
@@ -144,13 +133,13 @@ pub struct BuiltinBackendConfig {
     /// Exact JSON-pointer paths to sanitize. Empty means every string leaf.
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub target_paths: Vec<String>,
-    /// Regex pattern used when `action = "regex_replace"`.
+    /// Regex pattern used when `action = "regex_replace"` or `action = "redact"`.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub pattern: Option<String>,
     /// Built-in detector preset used when you do not want to write a regex.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub detector: Option<String>,
-    /// Replacement text used when `action = "regex_replace"`.
+    /// Replacement text used when `action = "regex_replace"` or `action = "redact"`.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub replacement: Option<String>,
     /// Masking token used when `action = "mask"`. Defaults to `*`.
@@ -231,7 +220,7 @@ crate::editor_config! {
         action => {
             label: "action",
             kind: Enum,
-            values: ["remove", "regex_replace", "hash", "mask"],
+            values: ["remove", "redact", "regex_replace", "hash", "mask"],
         },
         target_paths => { label: "target_paths", kind: Json },
         pattern => { label: "pattern", kind: String, optional: true },
@@ -335,7 +324,7 @@ fn builtin_action_schema(
 ) -> schemars::schema::Schema {
     string_enum_schema(
         generator,
-        &["remove", "regex_replace", "hash", "mask"],
+        &["remove", "redact", "regex_replace", "hash", "mask"],
         Some("remove"),
     )
 }
@@ -573,7 +562,7 @@ fn validate_builtin_action_requirements(
 
     if !matches!(
         builtin.action.as_str(),
-        "remove" | "regex_replace" | "hash" | "mask"
+        "remove" | "redact" | "regex_replace" | "hash" | "mask"
     ) {
         push_policy_diag(
             diagnostics,
@@ -581,11 +570,14 @@ fn validate_builtin_action_requirements(
             "pii_redaction.unsupported_value",
             Some(PII_REDACTION_PLUGIN_KIND.to_string()),
             Some("builtin.action".to_string()),
-            "builtin.action must be 'remove', 'regex_replace', 'hash', or 'mask'".to_string(),
+            "builtin.action must be 'remove', 'redact', 'regex_replace', 'hash', or 'mask'"
+                .to_string(),
         );
     }
 
-    if builtin.action == "regex_replace" && builtin.pattern.is_none() && builtin.detector.is_none()
+    if matches!(builtin.action.as_str(), "regex_replace" | "redact")
+        && builtin.pattern.is_none()
+        && builtin.detector.is_none()
     {
         push_policy_diag(
             diagnostics,
@@ -593,7 +585,7 @@ fn validate_builtin_action_requirements(
             "pii_redaction.unsupported_value",
             Some(PII_REDACTION_PLUGIN_KIND.to_string()),
             Some("builtin.pattern".to_string()),
-            "builtin.pattern or builtin.detector is required when builtin.action = 'regex_replace'"
+            "builtin.pattern or builtin.detector is required when builtin.action = 'regex_replace' or 'redact'"
                 .to_string(),
         );
     }
@@ -708,701 +700,6 @@ fn register_builtin_backend(
     }
 
     Ok(())
-}
-
-#[derive(Clone)]
-struct CompiledBuiltinBackend {
-    action: BuiltinAction,
-    target_paths: Arc<Vec<String>>,
-    codec: Option<Arc<dyn BuiltinRequestResponseCodec>>,
-    codec_name: Option<BuiltinCodecName>,
-}
-
-#[derive(Clone)]
-enum BuiltinAction {
-    Remove,
-    Hash {
-        matcher: Option<Arc<Regex>>,
-    },
-    Mask {
-        matcher: Option<Arc<Regex>>,
-        strategy: BuiltinMaskStrategy,
-    },
-    RegexReplace {
-        pattern: Arc<Regex>,
-        replacement: Arc<String>,
-    },
-}
-
-#[derive(Clone)]
-enum BuiltinMaskStrategy {
-    Generic {
-        mask_char: Arc<String>,
-        unmasked_prefix: usize,
-        unmasked_suffix: usize,
-    },
-    DetectorDefault {
-        detector: BuiltinDetector,
-        mask_char: Arc<String>,
-    },
-}
-
-#[derive(Clone, Copy)]
-enum BuiltinCodecName {
-    OpenAIChat,
-    OpenAIResponses,
-    AnthropicMessages,
-}
-
-trait BuiltinRequestResponseCodec: LlmCodec + LlmResponseCodec + Send + Sync {}
-
-impl<T> BuiltinRequestResponseCodec for T where T: LlmCodec + LlmResponseCodec + Send + Sync {}
-
-impl CompiledBuiltinBackend {
-    fn new(config: BuiltinBackendConfig, codec_name: Option<String>) -> PluginResult<Self> {
-        let detector = config
-            .detector
-            .as_deref()
-            .map(BuiltinDetector::parse)
-            .transpose()?;
-        let matcher = compile_builtin_matcher(config.pattern.clone(), detector)?;
-        let action = match config.action.as_str() {
-            "remove" => BuiltinAction::Remove,
-            "hash" => BuiltinAction::Hash { matcher },
-            "mask" => BuiltinAction::Mask {
-                matcher,
-                strategy: build_mask_strategy(&config, detector),
-            },
-            "regex_replace" => {
-                let pattern = matcher.ok_or_else(|| {
-                    PluginError::InvalidConfig(
-                        "builtin.pattern or builtin.detector is required when builtin.action = 'regex_replace'".to_string(),
-                    )
-                })?;
-                BuiltinAction::RegexReplace {
-                    pattern,
-                    replacement: Arc::new(
-                        config
-                            .replacement
-                            .unwrap_or_else(|| "[REDACTED]".to_string()),
-                    ),
-                }
-            }
-            other => {
-                return Err(PluginError::InvalidConfig(format!(
-                    "unsupported builtin.action '{other}'"
-                )));
-            }
-        };
-
-        Ok(Self {
-            action,
-            target_paths: Arc::new(config.target_paths),
-            codec_name: codec_name.as_deref().and_then(BuiltinCodecName::parse),
-            codec: codec_name
-                .as_deref()
-                .map(instantiate_builtin_codec)
-                .transpose()?,
-        })
-    }
-
-    fn sanitize_json_preorder_dfs(&self, value: Json) -> Json {
-        self.sanitize_json_preorder_dfs_at_path(value, &mut Vec::new())
-            .unwrap_or(Json::Null)
-    }
-
-    fn sanitize_json_preorder_dfs_at_path(
-        &self,
-        value: Json,
-        path_segments: &mut Vec<String>,
-    ) -> Option<Json> {
-        match value {
-            Json::String(text) => {
-                if self.matches_current_preorder_path(path_segments) {
-                    self.sanitize_string_value(text)
-                } else {
-                    Some(Json::String(text))
-                }
-            }
-            Json::Array(items) => Some(Json::Array(
-                items
-                    .into_iter()
-                    .enumerate()
-                    .map(|(index, item)| {
-                        path_segments.push(index.to_string());
-                        let sanitized = self
-                            .sanitize_json_preorder_dfs_at_path(item, path_segments)
-                            .unwrap_or(Json::Null);
-                        path_segments.pop();
-                        sanitized
-                    })
-                    .collect(),
-            )),
-            Json::Object(map) => Some(Json::Object(
-                map.into_iter()
-                    .filter_map(|(key, value)| {
-                        path_segments.push(escape_json_pointer_segment(&key));
-                        let sanitized =
-                            self.sanitize_json_preorder_dfs_at_path(value, path_segments);
-                        path_segments.pop();
-                        sanitized.map(|sanitized| (key, sanitized))
-                    })
-                    .collect(),
-            )),
-            other => {
-                if self.matches_current_preorder_path(path_segments)
-                    && matches!(self.action, BuiltinAction::Remove)
-                {
-                    None
-                } else {
-                    Some(other)
-                }
-            }
-        }
-    }
-
-    fn matches_current_preorder_path(&self, path_segments: &[String]) -> bool {
-        if self.target_paths.is_empty() {
-            return true;
-        }
-        let current_path = render_json_pointer_path(path_segments);
-        self.target_paths.iter().any(|path| path == &current_path)
-    }
-
-    fn sanitize_string_value(&self, text: String) -> Option<Json> {
-        match &self.action {
-            BuiltinAction::Remove => None,
-            BuiltinAction::Hash { matcher } => Some(Json::String(match matcher {
-                Some(matcher) => matcher
-                    .replace_all(&text, |captures: &regex::Captures<'_>| {
-                        hex_sha256(
-                            captures
-                                .get(0)
-                                .map(|capture| capture.as_str())
-                                .unwrap_or(""),
-                        )
-                    })
-                    .into_owned(),
-                None => hex_sha256(&text),
-            })),
-            BuiltinAction::Mask { matcher, strategy } => Some(Json::String(match matcher {
-                Some(matcher) => matcher
-                    .replace_all(&text, |captures: &regex::Captures<'_>| {
-                        mask_with_strategy(
-                            captures
-                                .get(0)
-                                .map(|capture| capture.as_str())
-                                .unwrap_or(""),
-                            strategy,
-                        )
-                    })
-                    .into_owned(),
-                None => mask_with_strategy(&text, strategy),
-            })),
-            BuiltinAction::RegexReplace {
-                pattern,
-                replacement,
-            } => Some(Json::String(
-                pattern
-                    .replace_all(&text, replacement.as_str())
-                    .into_owned(),
-            )),
-        }
-    }
-
-    fn sanitize_request_with_codec(&self, request: &LlmRequest) -> Option<LlmRequest> {
-        let codec = self.codec.as_ref()?;
-        let annotated = codec.decode(request).ok()?;
-        let sanitized_annotated = sanitize_serializable_with_backend(self, annotated).ok()?;
-        codec.encode(&sanitized_annotated, request).ok()
-    }
-
-    fn sanitize_response_with_codec(&self, payload: Json) -> Option<Json> {
-        let codec = self.codec.as_ref()?;
-        let codec_name = self.codec_name?;
-        let annotated = codec.decode_response(&payload).ok()?;
-        let sanitized_annotated = sanitize_serializable_with_backend(self, annotated).ok()?;
-        Some(codec_name.overlay_response_payload(payload, &sanitized_annotated))
-    }
-}
-
-fn tool_sanitize_callback(backend: CompiledBuiltinBackend) -> ToolSanitizeFn {
-    Arc::new(move |_name: &str, payload: Json| backend.sanitize_json_preorder_dfs(payload))
-}
-
-fn llm_sanitize_request_callback(backend: CompiledBuiltinBackend) -> LlmSanitizeRequestFn {
-    Arc::new(move |mut request: LlmRequest| {
-        if let Some(encoded) = backend.sanitize_request_with_codec(&request) {
-            return encoded;
-        }
-        request.content = backend.sanitize_json_preorder_dfs(request.content);
-        request
-    })
-}
-
-fn llm_sanitize_response_callback(backend: CompiledBuiltinBackend) -> LlmSanitizeResponseFn {
-    Arc::new(move |payload: Json| {
-        if backend.target_paths.is_empty() {
-            return backend.sanitize_json_preorder_dfs(payload);
-        }
-
-        let payload = backend
-            .sanitize_response_with_codec(payload.clone())
-            .unwrap_or(payload);
-        backend.sanitize_json_preorder_dfs(payload)
-    })
-}
-
-fn render_json_pointer_path(path_segments: &[String]) -> String {
-    if path_segments.is_empty() {
-        return String::new();
-    }
-    let mut rendered = String::new();
-    for segment in path_segments {
-        rendered.push('/');
-        rendered.push_str(segment);
-    }
-    rendered
-}
-
-fn escape_json_pointer_segment(segment: &str) -> String {
-    segment.replace('~', "~0").replace('/', "~1")
-}
-
-fn hex_sha256(text: &str) -> String {
-    let digest = Sha256::digest(text.as_bytes());
-    let mut output = String::with_capacity(digest.len() * 2);
-    for byte in digest {
-        use std::fmt::Write as _;
-        let _ = write!(&mut output, "{byte:02x}");
-    }
-    output
-}
-
-fn mask_text(
-    text: &str,
-    mask_char: &str,
-    unmasked_prefix: usize,
-    unmasked_suffix: usize,
-) -> String {
-    let chars: Vec<char> = text.chars().collect();
-    let len = chars.len();
-    if len <= unmasked_prefix + unmasked_suffix {
-        return text.to_string();
-    }
-
-    let mut output = String::new();
-    for ch in chars.iter().take(unmasked_prefix) {
-        output.push(*ch);
-    }
-    for _ in 0..(len - unmasked_prefix - unmasked_suffix) {
-        output.push_str(mask_char);
-    }
-    for ch in chars.iter().skip(len - unmasked_suffix) {
-        output.push(*ch);
-    }
-    output
-}
-
-fn build_mask_strategy(
-    config: &BuiltinBackendConfig,
-    detector: Option<BuiltinDetector>,
-) -> BuiltinMaskStrategy {
-    let mask_char = Arc::new(config.mask_char.clone().unwrap_or_else(|| "*".to_string()));
-    match detector {
-        Some(detector) if config.unmasked_prefix.is_none() && config.unmasked_suffix.is_none() => {
-            BuiltinMaskStrategy::DetectorDefault {
-                detector,
-                mask_char,
-            }
-        }
-        _ => BuiltinMaskStrategy::Generic {
-            mask_char,
-            unmasked_prefix: config.unmasked_prefix.unwrap_or(0),
-            unmasked_suffix: config.unmasked_suffix.unwrap_or(0),
-        },
-    }
-}
-
-fn mask_with_strategy(text: &str, strategy: &BuiltinMaskStrategy) -> String {
-    match strategy {
-        BuiltinMaskStrategy::Generic {
-            mask_char,
-            unmasked_prefix,
-            unmasked_suffix,
-        } => mask_text(text, mask_char.as_str(), *unmasked_prefix, *unmasked_suffix),
-        BuiltinMaskStrategy::DetectorDefault {
-            detector,
-            mask_char,
-        } => detector.default_mask(text, mask_char.as_str()),
-    }
-}
-
-fn compile_builtin_matcher(
-    pattern: Option<String>,
-    detector: Option<BuiltinDetector>,
-) -> PluginResult<Option<Arc<Regex>>> {
-    let pattern_text = match (pattern, detector) {
-        (Some(pattern), None) => Some(pattern),
-        (None, Some(detector)) => Some(detector.regex_pattern().to_string()),
-        (None, None) => None,
-        (Some(_), Some(_)) => {
-            return Err(PluginError::InvalidConfig(
-                "builtin.pattern and builtin.detector cannot both be set".to_string(),
-            ));
-        }
-    };
-
-    let Some(pattern_text) = pattern_text else {
-        return Ok(None);
-    };
-
-    let pattern = Regex::new(&pattern_text).map_err(|err| {
-        PluginError::InvalidConfig(format!(
-            "invalid builtin matcher regex '{pattern_text}': {err}"
-        ))
-    })?;
-    Ok(Some(Arc::new(pattern)))
-}
-fn instantiate_builtin_codec(
-    codec_name: &str,
-) -> PluginResult<Arc<dyn BuiltinRequestResponseCodec>> {
-    let codec: Arc<dyn BuiltinRequestResponseCodec> = match codec_name {
-        "openai_chat" => Arc::new(OpenAIChatCodec),
-        "openai_responses" => Arc::new(OpenAIResponsesCodec),
-        "anthropic_messages" => Arc::new(AnthropicMessagesCodec),
-        other => {
-            return Err(PluginError::InvalidConfig(format!(
-                "unsupported codec '{other}'"
-            )));
-        }
-    };
-    Ok(codec)
-}
-
-impl BuiltinCodecName {
-    fn parse(value: &str) -> Option<Self> {
-        match value {
-            "openai_chat" => Some(Self::OpenAIChat),
-            "openai_responses" => Some(Self::OpenAIResponses),
-            "anthropic_messages" => Some(Self::AnthropicMessages),
-            _ => None,
-        }
-    }
-
-    fn overlay_response_payload(self, payload: Json, annotated: &AnnotatedLlmResponse) -> Json {
-        match self {
-            Self::OpenAIChat => overlay_openai_chat_response(payload, annotated),
-            Self::OpenAIResponses => overlay_openai_responses_response(payload, annotated),
-            Self::AnthropicMessages => overlay_anthropic_response(payload, annotated),
-        }
-    }
-}
-
-fn overlay_openai_chat_response(mut payload: Json, annotated: &AnnotatedLlmResponse) -> Json {
-    let Some(root) = payload.as_object_mut() else {
-        return payload;
-    };
-    set_optional_string_field(root, "id", annotated.id.as_deref());
-    set_optional_string_field(root, "model", annotated.model.as_deref());
-
-    let Some(choice) = root
-        .get_mut("choices")
-        .and_then(Json::as_array_mut)
-        .and_then(|choices| choices.first_mut())
-        .and_then(Json::as_object_mut)
-    else {
-        return payload;
-    };
-
-    set_optional_string_field(
-        choice,
-        "finish_reason",
-        annotated
-            .finish_reason
-            .as_ref()
-            .map(openai_chat_finish_reason),
-    );
-
-    let Some(message) = choice.get_mut("message").and_then(Json::as_object_mut) else {
-        return payload;
-    };
-    set_optional_string_field(
-        message,
-        "content",
-        annotated_message_text(annotated.message.as_ref()).as_deref(),
-    );
-    overlay_openai_chat_tool_calls(message, annotated.tool_calls.as_deref());
-    payload
-}
-
-fn overlay_openai_responses_response(mut payload: Json, annotated: &AnnotatedLlmResponse) -> Json {
-    let Some(root) = payload.as_object_mut() else {
-        return payload;
-    };
-    set_optional_string_field(root, "id", annotated.id.as_deref());
-    set_optional_string_field(root, "model", annotated.model.as_deref());
-    set_optional_string_field(
-        root,
-        "status",
-        annotated
-            .finish_reason
-            .as_ref()
-            .map(openai_responses_status),
-    );
-
-    if let Some(items) = root.get_mut("output").and_then(Json::as_array_mut) {
-        overlay_output_text_blocks(items, annotated_message_text(annotated.message.as_ref()));
-        overlay_openai_responses_tool_calls(items, annotated.tool_calls.as_deref());
-    }
-    payload
-}
-
-fn overlay_anthropic_response(mut payload: Json, annotated: &AnnotatedLlmResponse) -> Json {
-    let Some(root) = payload.as_object_mut() else {
-        return payload;
-    };
-    set_optional_string_field(root, "id", annotated.id.as_deref());
-    set_optional_string_field(root, "model", annotated.model.as_deref());
-    set_optional_string_field(
-        root,
-        "stop_reason",
-        annotated.finish_reason.as_ref().map(anthropic_stop_reason),
-    );
-
-    if let Some(blocks) = root.get_mut("content").and_then(Json::as_array_mut) {
-        overlay_anthropic_text_blocks(blocks, annotated_message_text(annotated.message.as_ref()));
-        overlay_anthropic_tool_calls(blocks, annotated.tool_calls.as_deref());
-    }
-    payload
-}
-
-fn overlay_openai_chat_tool_calls(
-    message: &mut Map<String, Json>,
-    tool_calls: Option<&[ResponseToolCall]>,
-) {
-    let Some(raw_calls) = message.get_mut("tool_calls").and_then(Json::as_array_mut) else {
-        return;
-    };
-    let Some(tool_calls) = tool_calls else {
-        message.remove("tool_calls");
-        return;
-    };
-
-    for (raw_call, sanitized_call) in raw_calls.iter_mut().zip(tool_calls.iter()) {
-        let Some(raw_call) = raw_call.as_object_mut() else {
-            continue;
-        };
-        set_optional_string_field(raw_call, "id", Some(sanitized_call.id.as_str()));
-        let Some(function) = raw_call.get_mut("function").and_then(Json::as_object_mut) else {
-            continue;
-        };
-        set_optional_string_field(function, "name", Some(sanitized_call.name.as_str()));
-        set_optional_string_field(
-            function,
-            "arguments",
-            Some(json_string(&sanitized_call.arguments).as_str()),
-        );
-    }
-}
-
-fn overlay_openai_responses_tool_calls(
-    items: &mut [Json],
-    tool_calls: Option<&[ResponseToolCall]>,
-) {
-    let Some(tool_calls) = tool_calls else {
-        return;
-    };
-
-    let mut sanitized_calls = tool_calls.iter();
-    for item in items {
-        let Some(item_type) = item.get("type").and_then(Json::as_str) else {
-            continue;
-        };
-        if item_type != "function_call" {
-            continue;
-        }
-        let Some(raw_call) = item.as_object_mut() else {
-            continue;
-        };
-        let Some(sanitized_call) = sanitized_calls.next() else {
-            break;
-        };
-        set_optional_string_field(raw_call, "call_id", Some(sanitized_call.id.as_str()));
-        set_optional_string_field(raw_call, "name", Some(sanitized_call.name.as_str()));
-        set_optional_string_field(
-            raw_call,
-            "arguments",
-            Some(json_string(&sanitized_call.arguments).as_str()),
-        );
-    }
-}
-
-fn overlay_anthropic_tool_calls(blocks: &mut [Json], tool_calls: Option<&[ResponseToolCall]>) {
-    let Some(tool_calls) = tool_calls else {
-        return;
-    };
-
-    let mut sanitized_calls = tool_calls.iter();
-    for block in blocks {
-        let Some(block_type) = block.get("type").and_then(Json::as_str) else {
-            continue;
-        };
-        if block_type != "tool_use" {
-            continue;
-        }
-        let Some(raw_call) = block.as_object_mut() else {
-            continue;
-        };
-        let Some(sanitized_call) = sanitized_calls.next() else {
-            break;
-        };
-        set_optional_string_field(raw_call, "id", Some(sanitized_call.id.as_str()));
-        set_optional_string_field(raw_call, "name", Some(sanitized_call.name.as_str()));
-        raw_call.insert("input".into(), sanitized_call.arguments.clone());
-    }
-}
-
-fn overlay_output_text_blocks(items: &mut [Json], message_text: Option<String>) {
-    let text_items = items.iter_mut().filter_map(|item| {
-        (item.get("type").and_then(Json::as_str) == Some("message"))
-            .then_some(item.get_mut("content"))
-            .flatten()
-            .and_then(Json::as_array_mut)
-    });
-    let Some(text) = message_text else {
-        for content in text_items {
-            for block in content.iter_mut() {
-                if block.get("type").and_then(Json::as_str) == Some("output_text") {
-                    if let Some(block) = block.as_object_mut() {
-                        block.remove("text");
-                    }
-                }
-            }
-        }
-        return;
-    };
-
-    let parts: Vec<&str> = text.split('\n').collect();
-    for content in text_items {
-        let mut text_blocks = content.iter_mut().filter_map(|block| {
-            (block.get("type").and_then(Json::as_str) == Some("output_text"))
-                .then_some(block.as_object_mut())
-                .flatten()
-        });
-        for (index, block) in text_blocks.by_ref().enumerate() {
-            let part = parts
-                .get(index)
-                .copied()
-                .or_else(|| (index == 0).then_some(text.as_str()));
-            set_optional_string_field(block, "text", part);
-        }
-    }
-}
-
-fn overlay_anthropic_text_blocks(blocks: &mut [Json], message_text: Option<String>) {
-    let parts = message_text
-        .as_deref()
-        .map(|text| text.split('\n').collect::<Vec<_>>());
-    let mut text_block_index = 0usize;
-
-    for block in blocks {
-        if block.get("type").and_then(Json::as_str) != Some("text") {
-            continue;
-        }
-        let Some(block) = block.as_object_mut() else {
-            continue;
-        };
-        let part = parts
-            .as_ref()
-            .and_then(|parts| parts.get(text_block_index).copied())
-            .or_else(|| {
-                (text_block_index == 0)
-                    .then(|| message_text.as_deref())
-                    .flatten()
-            });
-        set_optional_string_field(block, "text", part);
-        text_block_index += 1;
-    }
-}
-
-fn annotated_message_text(message: Option<&MessageContent>) -> Option<String> {
-    match message? {
-        MessageContent::Text(text) => Some(text.clone()),
-        MessageContent::Parts(parts) => {
-            let text_parts: Vec<&str> = parts
-                .iter()
-                .filter_map(|part| match part {
-                    ContentPart::Text { text } => Some(text.as_str()),
-                    ContentPart::ImageUrl { .. } => None,
-                })
-                .collect();
-            (!text_parts.is_empty()).then(|| text_parts.join("\n"))
-        }
-    }
-}
-
-fn set_optional_string_field(object: &mut Map<String, Json>, key: &str, value: Option<&str>) {
-    match value {
-        Some(value) => {
-            object.insert(key.to_string(), Json::String(value.to_string()));
-        }
-        None => {
-            object.remove(key);
-        }
-    }
-}
-
-fn json_string(value: &Json) -> String {
-    serde_json::to_string(value).unwrap_or_else(|_| "null".to_string())
-}
-
-fn openai_chat_finish_reason(reason: &FinishReason) -> &str {
-    match reason {
-        FinishReason::Complete => "stop",
-        FinishReason::Length => "length",
-        FinishReason::ToolUse => "tool_calls",
-        FinishReason::ContentFilter => "content_filter",
-        FinishReason::Unknown(other) => other.as_str(),
-    }
-}
-
-fn openai_responses_status(reason: &FinishReason) -> &str {
-    match reason {
-        FinishReason::Complete => "completed",
-        FinishReason::Length | FinishReason::ContentFilter => "incomplete",
-        FinishReason::ToolUse => "completed",
-        FinishReason::Unknown(other) => other.as_str(),
-    }
-}
-
-fn anthropic_stop_reason(reason: &FinishReason) -> &str {
-    match reason {
-        FinishReason::Complete => "end_turn",
-        FinishReason::Length => "max_tokens",
-        FinishReason::ToolUse => "tool_use",
-        FinishReason::ContentFilter => "refusal",
-        FinishReason::Unknown(other) => other.as_str(),
-    }
-}
-
-fn sanitize_serializable_with_backend<T>(
-    backend: &CompiledBuiltinBackend,
-    value: T,
-) -> PluginResult<T>
-where
-    T: Serialize + DeserializeOwned,
-{
-    let value = serde_json::to_value(value).map_err(|err| {
-        PluginError::Internal(format!(
-            "failed to serialize value for PII redaction: {err}"
-        ))
-    })?;
-    serde_json::from_value(backend.sanitize_json_preorder_dfs(value)).map_err(|err| {
-        PluginError::Internal(format!(
-            "failed to deserialize sanitized value for PII redaction: {err}"
-        ))
-    })
 }
 
 fn validate_unknown_fields(
