+++ docs/pii-redaction-plugin/about.mdx
@@ -34,6 +34,33 @@ Start here when you need to:
 - Use built-in detector presets for common values such as emails, phone
   numbers, URLs, API keys, and IP addresses without writing custom regexes.
 
+## Plugin Versus Middleware
+
+`pii_redaction` is built on top of NeMo Relay's existing sanitize-guardrail
+middleware.
+
+That means this plugin does **not** introduce a separate runtime mechanism.
+Instead, it packages a common privacy policy behind a first-party config
+surface.
+
+Use the plugin when you want:
+
+- a reusable privacy policy that many applications or teams can share
+- declarative config through `plugins.toml` or `nemo-relay plugins edit`
+- built-in actions, detector presets, and codec-aware LLM handling
+- a supported, documented NeMo Relay surface instead of hand-registered callbacks
+
+Use raw sanitize-guardrail middleware when you want:
+
+- custom redaction logic authored directly in application code
+- dynamic behavior based on runtime state, external lookups, or one-off heuristics
+- experiments that are too app-specific to become a first-party plugin contract
+
+So the distinction is:
+
+- middleware is the **mechanism**
+- `pii_redaction` is the **packaged policy**
+
 ## Current Scope
 
 The built-in plugin currently exposes four managed sanitize surfaces:
@@ -43,9 +70,10 @@ The built-in plugin currently exposes four managed sanitize surfaces:
 - `tool_input`
 - `tool_output`
 
-The current built-in backend supports three actions:
+The current built-in backend supports five actions:
 
 - `remove`
+- `redact`
 - `regex_replace`
 - `hash`
 - `mask`
